pyproject.toml
[project]
name = "booted-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.121.1",
    "httpx>=0.28.1",
    "pillow>=11.3.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "roboflow>=1.2.11",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "ultralytics>=8.3.228",
    "uvicorn>=0.38.0",
]
